Load admin specified fonts for printing at runtime.
It would be nice if we could load the ttf directly instead
of needing it base64 encoded first.

I'm thinking all this font config stuff maybe should be
handled in printLayout.js instead of scattered all over.
Right now this is an issue with needing to provide regular
and bold versions of the typeface without any obvious
reason.  So maybe instead of sending in a props.font,
the user should just override the layout(s).

Passing in the FONTS is ugly.  Guessing the URL of the font
is ugly.  This should be cleaned up.  Probably pass in the
URL and the font family name in props and keep the rest of
the stuff internal.